#60 move dtos to separate module
