+++ backend/src/models/mod.rs
@@ -1,6 +1,6 @@
 //! Entities and DTOs used in `PermaplanT`.
 
-pub mod entity;
 pub mod dto;
+pub mod entity;
 pub mod r#enum;
 pub mod response;
