+++ backend/src/services/plants_service.rs
@@ -4,14 +4,15 @@ use actix_web::web::Data;
 
 use crate::{
     config::db::Pool,
-    error::ServiceError, models::{dto::plants::PlantsDTO, entity::Plants},
+    error::ServiceError,
+    models::{dto::PlantsDto, entity::Plants},
 };
 
 /// Fetch all plants from the database.
 ///
 /// # Errors
 /// If the connection to the database could not be established.
-pub fn find_all(pool: &Data<Pool>) -> Result<Vec<PlantsDTO>, ServiceError> {
+pub fn find_all(pool: &Data<Pool>) -> Result<Vec<PlantsDto>, ServiceError> {
     let mut conn = pool.get()?;
     let result = Plants::find_all(&mut conn)?;
     Ok(result)
